+++ pycantonese/word_segmentation.py
@@ -12,22 +12,53 @@ from pycantonese.data.rime_cantonese import (
 
 _MAX_WORD_LENGTH = 5
 
+_ALLOWED_WORDS = None
+_DISALLOWED_WORDS = None
+
+
+class Segmenter(LongestStringMatching):
+    """A customizable word segmentation model based on longest string match."""
+
+    def __init__(
+        self,
+        *,
+        max_word_length=_MAX_WORD_LENGTH,
+        allow=None,
+        disallow=None,
+    ):
+        """
+        Initialize a Segmenter object.
+
+        :param max_word_length: Maximal word length this model allows.
+        :param allow: An iterable of word strings to allow as words.
+        :param disallow: An iterable of word strings to disallow as words.
+        """
+        super(Segmenter, self).__init__(max_word_length=max_word_length)
+
+        # Train with HKCanCor data.
+        self.fit(hkcancor().sents())
+
+        # Train with rime-cantonese data.
+        self._words |= CHARS_TO_JYUTPING.keys()
+        self._words |= LETTERED.keys()
+        self._words |= set(PHRASES)
+        self._words |= set(MAPS)
+
+        # Adjust with the allowed and disallowed words.
+        self._words |= allow or set()
+        self._words -= disallow or set()
+
+    def _predict_sent(self, sent_str):
+        segmented = super(Segmenter, self)._predict_sent(sent_str)
+        return segmented
+
 
 @lru_cache(maxsize=1)
-def _get_word_segmentation_model():
-    model = LongestStringMatching(max_word_length=_MAX_WORD_LENGTH)
-    corpus = hkcancor()
-    # Train with HKCanCor data.
-    model.fit(corpus.sents())
-    # Train with rime-cantonese data.
-    model._words |= CHARS_TO_JYUTPING.keys()
-    model._words |= LETTERED.keys()
-    model._words |= set(PHRASES)
-    model._words |= set(MAPS)
-    return model
+def _get_default_segmenter():
+    return Segmenter()
 
 
-def segment(sent_str):
+def segment(sent_str, cls=None):
     """
     Segment the unsegmented sentence.
 
@@ -38,6 +69,10 @@ def segment(sent_str):
     characters.
 
     :param sent_str: Unsegmented sentence string
+    :param cls: A custom `Segmenter` class object for setting the maximal
+        word length (default = 5) and words to allow or disallow.
+        If not provided, a default segmenter is used, with maximal word
+        length = 5.
 
     :return: Segmented sentence
 
@@ -45,7 +80,9 @@ def segment(sent_str):
     """
     if not sent_str:
         return []
-    model = _get_word_segmentation_model()
-    # TODO: English words are parsed letter by letter, which is not good.
-    segmented = list(model.predict([sent_str]))[0]
+    if cls is None:
+        cls = _get_default_segmenter()
+    if type(cls) != Segmenter:
+        raise TypeError(f"`segmenter` must be a Segmenter object: {cls}")
+    segmented = list(cls.predict([sent_str]))[0]
     return segmented